collab_requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
#certifi==2024.6.2
cffi==1.16.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
#debugpy==1.8.1
#decorator==5.1.1
distro==1.9.0
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
#ipykernel==6.29.4
#ipython==8.25.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.4.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.31.0
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
protobuf==4.25.3
#psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
#python-dateutil==2.9.0.post0
python-dotenv==1.0.1
#pywin32==306
#pyzmq==26.0.3
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
stack-data==0.6.3
#tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
validators==0.28.3
wcwidth==0.2.13
